fix: main 앱을 db 앱으로 이동
